chore: move renderAriaTree to isomorphic
